fix: rename com.stremio.icons to com.stremio.icon
For consistency reasons only.
