+++ android/src/commonMain/kotlin/com/stremio/icon/StremioIcons.kt
@@ -1,3 +1,3 @@
-package com.stremio.icons
+package com.stremio.icon
 
 val StremioIcons = MR.images
